# docs/swagger.yaml
basePath: /api
definitions:
...
  job.Job:
    properties:
      created_at:
        type: string
      description:
        type: string
      expire_at:
        type: string
      hr_user_id:
        type: string
      id:
        type: string
      location:
        description: JSON string
        type: string
      requirements:
        description: JSON string
        type: string
      salary_range:
        description: JSON string
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  job.jobReq:
    properties:
      description:
        type: string
      location:
        type: string
      requirements:
        type: string
      salary_range:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  job.statusReq:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  resume.Resume:
    properties:
      attachment_urls:
        items:
          type: string
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  resume.resumeReq:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/maxfeizi04-cloud/recruitment-platform
  description: 放心招聘平台后端接口文档，支持求职者、招聘者、管理员三端
  title: 放心招聘平台 API
  version: "2.0"
paths:
...
  /jobs:
    get:
      consumes:
      - application/json
      parameters:
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取职位列表
      tags:
      - 职位
    post:
      consumes:
      - application/json
      parameters:
      - description: 职位信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/job.jobReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.Job'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 发布职位
      tags:
      - 职位
  /jobs/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 职位ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/job.Job'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取职位详情
      tags:
      - 职位
    put:
      consumes:
      - application/json
      parameters:
      - description: 职位ID
        in: path
        name: id
        required: true
        type: string
      - description: 职位信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/job.jobReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新职位
      tags:
      - 职位
  /jobs/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: 职位ID
        in: path
        name: id
        required: true
        type: string
      - description: 状态值（active/paused/closed）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/job.statusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新职位状态
      tags:
      - 职位
  /jobs/my:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取我发布的职位
      tags:
      - 职位
  /jobs/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: 搜索关键词
        in: query
        name: q
        type: string
      - description: 城市
        in: query
        name: city
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 搜索职位
      tags:
      - 职位
...
  /recommend/candidates:
    get:
      consumes:
      - application/json
      parameters:
      - description: 职位ID
        in: query
        name: job_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 推荐候选人
      tags:
      - 推荐
  /recommend/jobs:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 推荐职位
      tags:
      - 推荐
  /resumes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取我的简历列表
      tags:
      - 简历
    post:
      consumes:
      - application/json
      parameters:
      - description: 简历标题和内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resume.resumeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resume.Resume'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建简历
      tags:
      - 简历
  /resumes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 简历ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除简历
      tags:
      - 简历
    put:
      consumes:
      - application/json
      parameters:
      - description: 简历ID
        in: path
        name: id
        required: true
        type: string
      - description: 简历标题和内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resume.resumeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新简历
      tags:
      - 简历
  /resumes/{id}/set-default:
    post:
      consumes:
      - application/json
      parameters:
      - description: 简历ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 设为默认简历
      tags:
      - 简历
  /resumes/{id}/upload-attachment:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 简历ID
        in: path
        name: id
        required: true
        type: string
      - description: 附件文件（PDF/DOC/DOCX/JPG/PNG，最大20MB）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 上传简历附件
      tags:
      - 简历
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"